adding caching for access links
HeartBeats now have a list of certificates, and digest hash of each access link. This allows the server to:

(1) without getting the complete acces link, will know if an access link is updated. This helps the app use its local copy, if it's still valid.

(2) server can audit and instruct home device if there are left-over fiends or missing ones.

For 1, see http://src.we-pn.com/mobile_app/issues/17/cache-access-link-and-hash-for-each-friend