add backwards compatibility for bullsye devices
On bookworm, kernel 6.6.x is removing older method to control GPIO.
However, the new method (lgpio, etc.) cannot be installed on bullseye
as is. There dependency graph gets too complex.

To solve this, for now we will detect this "legacy" GPIO, and use the old method.
More testing is required to make sure both paths are working.